1. compile error related to program_option_lite is fixed 2. Update byteStreamToSampleStream, sampleStreamToByteStream    1) removing zeroCount for emulation prevention bytes    2) changing the size of startcode : 4bytes for VPS, SPS, PPS and the first NAL of new frames

+++ source/lib/PccLibEncoder/source/PCCNormalsGenerator.cpp
@@ -282,10 +282,10 @@ void PCCNormalsGenerator3::orientNormals( const PCCPointSet3&                   
       PCCVector3D( 0, 1, 0 ), 
       PCCVector3D( 0, 0, 1 ), 
     };
-    int planeWidth[3] = { (bb.max_[1] - bb.min_[1])+1,
+    double planeWidth[3] = { (bb.max_[1] - bb.min_[1])+1,
       (bb.max_[2] - bb.min_[2])+1,
       (bb.max_[0] - bb.min_[0])+1};
-    int planeHeight[3] = {(bb.max_[2]-bb.min_[2])+1, 
+    double planeHeight[3] = {(bb.max_[2]-bb.min_[2])+1, 
       (bb.max_[0]-bb.min_[0])+1, 
       (bb.max_[1]-bb.min_[1])+1};
     //allocating projection planes
